requirements.txt
langchain
openai
python-dotenv
langchain-community
streamlit
langchain_openai